refactor: share receiver level normalizers with command dispatch
