+++ src/rigplane/core/command_service.py
@@ -15,6 +15,10 @@ from dataclasses import dataclass, replace
 from typing import TYPE_CHECKING, Any, Protocol
 
 from rigplane.core.acquisition_scheduler import AcquisitionPriority, AcquisitionStatus
+from rigplane.core.command_dispatch import (
+    _af_level_from_param as _af_level_from_param,
+    _raw_int_level_from_param as _raw_int_level_from_param,
+)
 from rigplane.core.exceptions import TimeoutError as RigplaneTimeoutError
 from rigplane.core.state_pipeline_contracts import (
     ChangeSet,
@@ -877,61 +881,6 @@ def _expected_value_for_path(intent: CommandIntent, path: FieldPath) -> Any:
 
 def _should_normalize_level_expectation(name: str, path: FieldPath) -> bool:
     return _NORMALIZED_LEVEL_EXPECTATION_COMMANDS.get(name) == path.name
-
-
-def _raw_int_level_from_param(value: Any) -> int:
-    """Coerce a raw-only level command param (MOR-1579).
-
-    ``set_rf_gain``/``set_sql``/``set_squelch``: both the web frontend
-    (``radio-intents.ts`` declares ``'integer'``) and the documented
-    HTTP/WS command catalog agree the wire value is always a raw 0-255
-    integer, never a normalized float. Dispatch on the JSON *type*, not
-    magnitude — a value in ``[0, 1]`` used to be silently reinterpreted as
-    normalized (MOR-1579's headline bug: raw level ``1`` became raw
-    ``255``). A non-int or an out-of-range int is a caller bug, not an
-    alternate encoding, so it raises instead of being coerced.
-
-    This value feeds both the StateStore readback expectation (via
-    :func:`_expected_value_for_path`) *and*, on the ``public_api`` sync
-    ingress (:mod:`rigplane.runtime.sync`), the actual value sent to the
-    radio (``_SyncCommandExecutor`` reads ``intent.params["squelch"]``
-    directly) — so this function is the actuation path there, not just
-    bookkeeping.
-    """
-    if isinstance(value, bool) or not isinstance(value, int):
-        raise ValueError(
-            f"level {value!r} must be a raw integer 0-255, not {type(value).__name__}"
-        )
-    if not (0 <= value <= 255):
-        raise ValueError(f"level {value!r} is out of the raw 0-255 domain")
-    return int(value)
-
-
-def _af_level_from_param(value: Any) -> int:
-    """Coerce ``set_af_level``'s type-dispatched level param (MOR-1579).
-
-    Two documented wire contracts coexist for this one intent: the
-    HTTP/WS command catalog (``docs/api/command-catalog.md``) declares
-    ``level: int`` on the raw 0-255 scale (see the live-hardware
-    validation recipe's ``level:35`` example, which expects raw BCD
-    ``0035``); the web frontend (``radio-intents.ts`` declares
-    ``'normalized'``) sends a JSON float in 0.0-1.0. Dispatch on JSON
-    type, never magnitude: an int is always raw, a float is always
-    normalized, matching MOR-334's original coercion for float input
-    while restoring int input to a true no-op. Out-of-domain values for
-    either type raise rather than being reinterpreted as the other.
-    """
-    if isinstance(value, bool):
-        raise ValueError(f"level {value!r} must be an int or a normalized float")
-    if isinstance(value, int):
-        if not (0 <= value <= 255):
-            raise ValueError(f"level {value!r} is out of the raw 0-255 domain")
-        return value
-    if isinstance(value, float):
-        if not (0.0 <= value <= 1.0):
-            raise ValueError(f"level {value!r} is out of the normalized 0.0-1.0 domain")
-        return max(0, min(255, round(value * 255)))
-    raise ValueError(f"level {value!r} must be an int or a normalized float")
 
 
 def _power_level_expectation_from_param(
